change Base to ValidatedSQLModel
ref #24
